Remove cleanSubVer in debug log

(cherry picked from commit cd4a25f)

+++ src/spork.cpp
@@ -96,9 +96,8 @@ bool IsSporkActive(int nSporkID) {
 }
 
 // grab the value of the spork on the network, or the default
-long GetSporkValue(int nSporkID) {
-    long r = 0;
-
+int GetSporkValue(int nSporkID) {
+    int r = 0;
     if (mapSporksActive.count(nSporkID)) {
         r = mapSporksActive[nSporkID].nValue;
     } else {
